Move get_person_location to case utils
