+++ custom/enikshay/case_utils.py
@@ -1,10 +1,17 @@
 import pytz
+from collections import namedtuple
 from django.utils.dateparse import parse_datetime
+
+from corehq.apps.locations.models import SQLLocation
 
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.util import post_case_blocks
 from corehq.form_processor.interfaces.dbaccessors import CaseAccessors
-from custom.enikshay.exceptions import ENikshayCaseNotFound
+from custom.enikshay.exceptions import (
+    ENikshayCaseNotFound,
+    NikshayCodeNotFound,
+    NikshayLocationNotFound,
+)
 from corehq.form_processor.exceptions import CaseNotFound
 
 CASE_TYPE_ADHERENCE = "adherence"
@@ -144,3 +151,32 @@ def update_case(domain, case_id, updated_properties, external_id=None):
         [CaseBlock(**kwargs).as_xml()],
         {'domain': domain}
     )
+
+
+def get_person_locations(person_case):
+    PersonLocationHierarchy = namedtuple('PersonLocationHierarchy', 'sto dto tu phi')
+    try:
+        phi_location = SQLLocation.objects.get(location_id=person_case.owner_id)
+    except SQLLocation.DoesNotExist:
+        raise NikshayLocationNotFound(
+            "Location with id {location_id} not found. This is the owner for person with id: {person_id}"
+            .format(location_id=person_case.owner_id, person_id=person_case.case_id)
+        )
+
+    try:
+        tu_location = phi_location.parent
+        district_location = tu_location.parent
+        city_location = district_location.parent
+        state_location = city_location.parent
+    except AttributeError:
+        raise NikshayLocationNotFound("Location structure error for person: {}".format(person_case.case_id))
+    try:
+        # TODO: verify how location codes will be stored
+        return PersonLocationHierarchy(
+            sto=state_location.metadata['nikshay_code'],
+            dto=district_location.metadata['nikshay_code'],
+            tu=tu_location.metadata['nikshay_code'],
+            phi=phi_location.metadata['nikshay_code'],
+        )
+    except (KeyError, AttributeError) as e:
+        raise NikshayCodeNotFound("Nikshay codes not found: {}".format(e))
