+++ custom/enikshay/exceptions.py
@@ -2,5 +2,13 @@ class ENikshayException(Exception):
     pass
 
 
-class ENikshayCaseNotFound(Exception):
+class ENikshayCaseNotFound(ENikshayException):
+    pass
+
+
+class NikshayLocationNotFound(ENikshayException):
+    pass
+
+
+class NikshayCodeNotFound(ENikshayException):
     pass
