+++ custom/enikshay/tests/test_case_utils.py
@@ -4,7 +4,8 @@ import pytz
 from django.test import TestCase
 
 from corehq.form_processor.interfaces.dbaccessors import CaseAccessors
-from custom.enikshay.tests.utils import ENikshayCaseStructureMixin
+from custom.enikshay.tests.utils import ENikshayCaseStructureMixin, ENikshayLocationStructureMixin
+from custom.enikshay.exceptions import NikshayLocationNotFound
 from corehq.form_processor.tests.utils import run_with_all_backends, FormProcessorTestUtils
 
 from custom.enikshay.case_utils import (
@@ -16,6 +17,7 @@ from custom.enikshay.case_utils import (
     update_case,
     get_open_occurrence_case_from_person,
     get_open_episode_case_from_occurrence,
+    get_person_locations,
 )
 
 
@@ -124,3 +126,26 @@ class ENikshayCaseUtilsTests(ENikshayCaseStructureMixin, TestCase):
             get_open_episode_case_from_occurrence(self.domain, self.occurrence_id).case_id,
             self.episode_id
         )
+
+
+class TestGetPersonLocations(ENikshayCaseStructureMixin, ENikshayLocationStructureMixin, TestCase):
+    def setUp(self):
+        super(TestGetPersonLocations, self).setUp()
+        self.cases = self.create_case_structure()
+
+    def test_get_person_locations(self):
+        self.assign_person_to_location(self.phi.location_id)
+        person_case = CaseAccessors(self.domain).get_case(self.person_id)
+        expected_locations = {
+            'sto': self.sto.metadata['nikshay_code'],
+            'dto': self.dto.metadata['nikshay_code'],
+            'tu': self.tu.metadata['nikshay_code'],
+            'phi': self.phi.metadata['nikshay_code'],
+        }
+        self.assertEqual(expected_locations, get_person_locations(person_case)._asdict())
+
+    def test_nikshay_location_not_found(self):
+        self.assign_person_to_location("-")
+        person_case = CaseAccessors(self.domain).get_case(self.person_id)
+        with self.assertRaises(NikshayLocationNotFound):
+            get_person_locations(person_case)
